requirements.txt
streamlit==1.50.0
pandas==2.3.3
numpy==2.0.2
praw==7.8.1
sentence-transformers==5.1.2
langchain==0.3.28
langchain-huggingface==0.3.1
faiss-cpu==1.13.0
